adding check for files index in dir array

+++ modules/base/classes/fileCache.php
@@ -199,9 +199,10 @@ class owa_fileCache extends owa_cache {
 		foreach ($tld['dirs'] as $k => $dir) {
 			
 			$sld = $this->readDir($dir);
-			$this->debug("Reading cache file list from: ". $dir);	
-			$this->deleteFiles($sld['files']);
-		
+			$this->debug("Reading cache file list from: ". $dir);
+			if (array_key_exists('files', $sld)) {	
+				$this->deleteFiles($sld['files']);
+			}
 			foreach ($sld['dirs'] as $sk => $sdir) {
 				$ssld = $this->readDir($sdir);
 				$this->debug("Reading cache file list from: ". $sdir);
